+++ OzzTradeDiary.NET/OzzTradeDiary/Models/Symbol.cs
@@ -74,6 +74,7 @@ namespace TD.Models
         /// The Id value of exchange record that related to the symbol.
         /// </summary>
         [Required(ErrorMessageResourceType = typeof(ErrorStrings), ErrorMessageResourceName = "Required")]
+        [Range(1, int.MaxValue, ErrorMessageResourceType = typeof(ErrorStrings), ErrorMessageResourceName ="ValueMin")]
         [Display(ResourceType = typeof(LocalizedStrings), Name = "ExchangeId")]
         public int ExchangeId { get; set; }